FIX: Refactor link code from main canvas file
